Add specific exit codes for each task failure type
This makes it so that I can deduce which tasks failed from the exit code
on the Task Scheduler view. I imagine this can be useful if the email
fails to send for some reason, and communicating results through exit
codes might be nice to future DevOps things in general. The constants
are stored in config.py and are easily extendable.