Draft ABC for generic updater
This makes each updater conform to a uniform interface. Each updater for
some social media platform just needs to specify how it parses the
global JSON configuration data and how it performs the updating. The
specific platform is transparent to the driver code, which then
instantiates an updater and runs interface methods.