chore: update task preview to un-nest renderables
Previously, the format_generic_task_preview helper would always nest the
given body into a Panel. This is desirable if it's a plain string (or
None i.e. not given, in which case a default string is used), but for
full-fledged renderables prepared by the caller such as tables, it is an
unnecessary level of nesting. This change makes the helper use the
renderable itself as "panel" if it's not a string.