feat: add option to specify date to update to
This is similar to the argument to the --dry-run option, except this
argument would be used for actual updating.

The date is supplied as an optional positional argument, as an ISO 8601
date string. It defaults to today if omitted, making it backwards
compatible. One application behind this feature is to manually run the
counters for the upcoming day ahead of time (such as when it is not
guaranteed the machine will be on until midnight and thus the task
scheduler will not run.)